requirements.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
atari-py==0.2.6
Box2D==2.3.10
certifi==2020.6.20
cffi==1.14.2
click==7.1.2
cloudpickle==1.2.2
colorama==0.4.3
cycler==0.10.0
cython==0.29.21
dill==0.3.2
docopt==0.6.2
future==0.18.2
git+https://github.com/pimdh/baselines.git@tweaks#egg=baselines
git+https://github.com/pimdh/hyperject@master
gitdb==4.0.5
GitPython==3.1.7
glfw==1.12.0
google-pasta==0.2.0
gtimer==1.0.0b5
gym==0.15.3
h5py==2.10.0
imageio==2.9.0
importlib-metadata==2.0.0
joblib==0.16.0
jsonpickle==1.4.1
kiwisolver==1.2.0
lockfile==0.12.2
Markdown==3.3.1
matplotlib==3.2.2
mkl-fft==1.0.6
mkl-random==1.0.1
mpi4py==3.0.3
mujoco-py==1.50.1.68
munch==2.5.0
numpy==1.16.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
Pillow==7.2.0
progressbar2==3.52.1
protobuf==3.13.0
py-cpuinfo==7.0.0
pycparser==2.20
pygame==2.0.0
pyglet==1.3.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
pytorch-ignite==0.4.2
pytz==2020.1
PyYAML==5.3.1
sacred==0.8.1
scikit-learn==0.23.2
scipy==1.1.0
six==1.15.0
smmap==3.0.4
stable-baselines==2.10.1
tb-nightly==1.14.0a20190301
tensorboard==1.14.0
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.48.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
